Execute cpu ops sequentially (#51)
Changes introduced in this PR:
- renamed message defines for CPU operations so they follow a pattern: __cudampi__CPU\<OPERATIONTYPE\>[REQ/RESP] (e.g. `__cudampi__CPUMALLOCRESP`) 
- Implemented __cudampi__cpuMemcpyAsync
- Introduced asynchronous execution for CPU operations
- Added print_double_array funtion to app-streams-collatz, so that results of running application in different configurations (with / without CPU) could be compared with simple `diff result_before.txt result_after.txt`
- Changed logging in slave code to use logger instead of printf 

**Summary of asynchronous execution for CPU operations:**

A task queue was introduced, where `dep_node_t` type implements queue node. Each task to be executed asynchronously is added to the queue with `allocateCpuTask` function and queue processing is done with `cpuTaskLauncher`. Task order is enforced with OpenMP task dependency mechanism, where each task in the queue depends on previous task, which ensures sequential execution. Task launching and freeing already executed nodes is done in `scheduleCpuTask` (called inside `cpuTaskLauncher` for each task) .
 
Since OpenMP tasks are tied to the thread that started them, in current implementation (where only one thread is processing CPU requests) when task is launched it will be pending execution until taskwait is called, because currently executing task (parallel region that processes incoming MPI messages) will have higher priority - after processing a message and adding it to queue it will proceed to wait for next message in `MPI_Probe`. To handle this behavior a following mechanism was implemented:

Async operations are added to the queue and there's additional CPU thread for executing these operations. This approach provides fully asynchronous way to execute CPU operations (just like GPU ones).  When synchronous operation is called, slave waits for all scheduled operations in the queue to complete. Additionally, tasks that should be processed asynchronously on master side (memcpy with deviceToHost flag) are being asynchronously received on master side with IRecv and added to the queue for further processing during synchronization. During deviceSynchronize call on master side, the code waits for all previously scheduled IRecvs to complete.  
---------

Co-authored-by: Piotr Sieński <s184297@student.pg.edu.pl>
Co-authored-by: psienski <piotr.sienski@intel.com>